REMOVED: nl_time* define that is not necessary FIXED: some warning about size_t uint32 conversion CLEANED: the types_nl by removing unused macro and put the # in the good position (must be the 1 character of the line)
git-svn-id: svn://svn.code.sf.net/p/nel/code/trunk@217 bf685bcc-7839-4fcd-9254-36823579bede

+++ nel/include/nel/misc/stream_inline.h
@@ -284,7 +284,7 @@ inline	void		IStream::serial(std::string &b)
 	}
 	else
 	{
-		len= b.size();
+		len= uint32(b.size());
 		if (len>1000000)
 			throw NLMISC::EInvalidDataStream( "IStream: Trying to write a string of %u bytes", len );
 		serial(len);
@@ -317,7 +317,7 @@ inline	void		IStream::serial(ucstring &b)
 	}
 	else
 	{
-		len= b.size();
+		len= uint32(b.size());
 		if (len>1000000)
 			throw NLMISC::EInvalidDataStream( "IStream: Trying to write an ucstring of %u bytes", len );
 		serial(len);
